+++ R/Aaaa.R
@@ -1,8 +1,10 @@
 ##' @importFrom grDevices pdf dev.off graphics.off
-##' @importFrom ggplot2 theme margin
+##' @importFrom ggplot2 theme margin ggplotGrob
 ##' @importFrom gridExtra grid.arrange arrangeGrob
 ##' @importFrom grid grid.draw gpar grid.newpage
 ##' @importFrom grid textGrob gList
+##' @importFrom assertthat assert_that
+##' @importFrom rlang flatten flatten_if
 ##'
 NULL